Alter Robolectric usage to finish AndroidX migration
Robolectric deprecated some APIs in preference to their new
AndroidX test library support:
http://robolectric.org/androidx_test/

This removes the usage of the deprecated APIs and implements the
necessary project changes to use the AndroidX style

- Activities must have <activity> AndroidManifest configuration, so
  our test-only Activity subclasses have to be public, and declared in
  a mergeable location (debug build variant for merge here)
- Robolectric test runner gone in favor of JUnit Runner
