+++ AnkiDroid/build.gradle
@@ -60,8 +60,6 @@ android {
     useLibrary 'org.apache.http.legacy'
 
     testOptions {
-        // we should use orchestrator, but app data shared on /sdcard makes it unstable
-        //execution 'ANDROIDX_TEST_ORCHESTRATOR'
         animationsDisabled true
 
         unitTests {
@@ -93,6 +91,11 @@ android {
         maxParallelForks = gradleTestMaxParallelForks
         systemProperties['junit.jupiter.execution.parallel.enabled'] = true
         systemProperties['junit.jupiter.execution.parallel.mode.default'] = "concurrent"
+    }
+    sourceSets {
+        debug {
+            manifest.srcFile 'src/test/AndroidManifest.xml'
+        }
     }
 }
 
@@ -196,15 +199,14 @@ dependencies {
     testImplementation 'org.powermock:powermock-api-mockito2:2.0.0'
     testImplementation 'org.hamcrest:hamcrest-all:1.3'
     testImplementation 'net.lachlanmckee:timber-junit-rule:1.0.1'
-    testImplementation 'androidx.test:core:1.1.0'
     testImplementation "org.robolectric:robolectric:4.2"
+    testImplementation 'androidx.test:core:1.1.0'
+    testImplementation 'androidx.test.ext:junit:1.1.0'
 
     // May need a resolution strategy for support libs to our versions
     androidTestImplementation 'com.azimolabs.conditionwatcher:conditionwatcher:0.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.1.1'
     androidTestImplementation 'androidx.test.ext:junit:1.1.0'
-    androidTestImplementation 'androidx.test:runner:1.1.1'
     androidTestImplementation 'androidx.test:rules:1.1.1'
-    androidTestUtil 'androidx.test:orchestrator:1.1.1'
 }
