+++ AnkiDroid/src/test/AndroidManifest.xml
@@ -1,0 +1,16 @@
+<manifest xmlns:tools="http://schemas.android.com/tools"
+          xmlns:android="http://schemas.android.com/apk/res/android">
+    <application tools:ignore="AllowBackup,GoogleAppIndexingWarning">
+
+        <!-- To use AndroidX Test APIs on Activity sub-classes, you must have valid
+             Activity entries in your AndroidManifest. Manifest merger only happens with
+             build variants, and tests always run with the debug build variant, so test-only
+             Activity subclasses are defined here, and this file is included in debug.sourceSet
+             https://developer.android.com/studio/build/manifest-merge
+             https://issuetracker.google.com/issues/37001185
+             -->
+        <activity
+            android:name="com.ichi2.anki.NullCollectionReviewer">
+        </activity>
+    </application>
+</manifest>
