+++ AnkiDroid/src/test/java/com/ichi2/anki/DeckPickerTest.java
@@ -2,12 +2,12 @@ package com.ichi2.anki;
 
 import android.content.Context;
 
+import androidx.test.core.app.ActivityScenario;
 import androidx.test.core.app.ApplicationProvider;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.robolectric.Robolectric;
-import org.robolectric.RobolectricTestRunner;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -15,7 +15,7 @@ import java.util.Map;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 
-@RunWith(RobolectricTestRunner.class)
+@RunWith(AndroidJUnit4.class)
 public class DeckPickerTest extends RobolectricTest {
 
     @Test
@@ -32,19 +32,25 @@ public class DeckPickerTest extends RobolectricTest {
         mCodeResponsePairs.put(503, context.getString(R.string.sync_too_busy));
         mCodeResponsePairs.put(504, context.getString(R.string.sync_error_504_gateway_timeout));
 
-        DeckPicker deckPicker = Robolectric.setupActivity(DeckPicker.class);
-        for (Map.Entry<Integer, String> entry : mCodeResponsePairs.entrySet()) {
-            assertEquals(deckPicker.rewriteError(entry.getKey()), entry.getValue());
+        try (ActivityScenario<DeckPicker> scenario = ActivityScenario.launch(DeckPicker.class)) {
+            scenario.onActivity(deckPicker -> {
+                for (Map.Entry<Integer, String> entry : mCodeResponsePairs.entrySet()) {
+                    assertEquals(deckPicker.rewriteError(entry.getKey()), entry.getValue());
+                }
+            });
         }
     }
 
     @Test
     public void verifyBadCodesNoMessage() {
-        DeckPicker deckPicker = Robolectric.setupActivity(DeckPicker.class);
-        assertNull(deckPicker.rewriteError(0));
-        assertNull(deckPicker.rewriteError(-1));
-        assertNull(deckPicker.rewriteError(1));
-        assertNull(deckPicker.rewriteError(Integer.MIN_VALUE));
-        assertNull(deckPicker.rewriteError(Integer.MAX_VALUE));
+        try (ActivityScenario<DeckPicker> scenario = ActivityScenario.launch(DeckPicker.class)) {
+            scenario.onActivity(deckPicker -> {
+                assertNull(deckPicker.rewriteError(0));
+                assertNull(deckPicker.rewriteError(-1));
+                assertNull(deckPicker.rewriteError(1));
+                assertNull(deckPicker.rewriteError(Integer.MIN_VALUE));
+                assertNull(deckPicker.rewriteError(Integer.MAX_VALUE));
+            });
+        }
     }
 }
