+++ AnkiDroid/src/test/java/com/ichi2/anki/ReviewerTest.java
@@ -5,38 +5,28 @@ import com.ichi2.libanki.Collection;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.Robolectric;
-import org.robolectric.RobolectricTestRunner;
+
+import androidx.test.core.app.ActivityScenario;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 
-@RunWith(RobolectricTestRunner.class)
+@RunWith(AndroidJUnit4.class)
 public class ReviewerTest extends RobolectricTest {
 
     @Test
     public void verifyStartupNoCollection() {
-        Reviewer reviewer = Robolectric.setupActivity(NullCollectionReviewer.class);
-        assertNull("Collection should have been null", reviewer.getCol());
+        try (ActivityScenario<NullCollectionReviewer> scenario = ActivityScenario.launch(NullCollectionReviewer.class)) {
+            scenario.onActivity(reviewer -> assertNull("Collection should have been null", reviewer.getCol()));
+        }
     }
 
     @Test
     public void verifyNormalStartup() {
-        Reviewer reviewer = Robolectric.setupActivity(Reviewer.class);
-        assertNotNull("Collection should be non-null", reviewer.getCol());
+        try (ActivityScenario<Reviewer> scenario = ActivityScenario.launch(Reviewer.class)) {
+            scenario.onActivity(reviewer -> assertNotNull("Collection should be non-null", reviewer.getCol()));
+        }
     }
 }
 
-class NullCollectionReviewer extends Reviewer {
-    @Override
-    public Collection getCol() {
-        return null;
-    }
-    @Override
-    public boolean colIsOpen() {
-        return false;
-    }
-    @Override
-    protected void onCollectionLoaded(Collection col) {
-        // it's not fair to expect the classes under test to handle this when we return null for getCol()
-    }
-}
